==> 1_Web_Scraping.ipynb <==
# Introduction

This program is created to retrieve information from the website to create a table so that the data can be analyzed.

# Web Scraping

# The formula for importing libraries required for web scraping
from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.by import By
import pandas as pd
import time
from scipy import stats
import numpy as np
url="https://www.tokopedia.com/search?navsource=&page=1&q=seblak&srp_component_id=02.01.00.00&srp_page_id=&srp_page_title=&st="


# The formula for doing web scraping from the tokopedia website
list_tokped= []
for i in range(1,11):
    url="https://www.tokopedia.com/search?navsource=&page={}&q=seblak&srp_component_id=02.01.00.00&srp_page_id=&srp_page_title=&st=".format(i)
    driver = webdriver.Chrome()
    driver.get(url)
    html = driver.page_source
    soup = BeautifulSoup(html, "html.parser")
    for item in soup.findAll('div',{"class":"css-1asz3by"}):
          Name = item.find('div',{"class":"prd_link-product-name css-3um8ox"}).text
          Price = item.find('div',{"class":"prd_link-product-price css-h66vau"}).text

          item3 = item.findAll('span', {"class":"prd_label-integrity css-1sgek4h"})
          if len(item3) > 0:
               Product_sold = item.find('span', {"class":"prd_label-integrity css-1sgek4h"}).text
          else:
               Product_sold = None
          item4 = item.findAll ('span', {"class":"prd_rating-average-text css-t70v7i"})
          if len(item4) > 0:
               Product_rating = item.find('span', {"class":"prd_rating-average-text css-t70v7i"}).text
          else:
               Product_rating = None
          for item2 in item.findAll('div',{"class":"css-1rn0irl"}):
               Seller = item2.findAll('span',{"class":"prd_link-shop-name css-1kdc32b flip"})[0].text
               Cities = item2.findAll('span', {"class":"prd_link-shop-loc css-1kdc32b flip"})[0].text
               list_tokped.append((Name, Price, Seller, Cities, Product_sold,Product_rating))

df = pd.DataFrame(list_tokped, columns=['Name','Price','Seller','Cities','Product sold', 'Product rating'])
df = df.drop_duplicates()
df

#The formula to export data into a csv file
df.to_csv('Seblak_Tokped.csv')
