==> 2_Data_Exploration_&_Cleaning.ipynb <==
# Data Exploration

This part focuses on the process of initial Data Exploration and Data Cleaning

# Import Libraries
import pandas as pd

#The formula for reading back a file that was previously exported to csv
df = pd.read_csv('Seblak_Tokped.csv')

#The formula for outputting information from the database
df.info()

Based on the results of the data reading above, there is information that all data types are objects, so they must be converted to the correct data type in order to be processed.

# Formula to display columns in the dataframe
df.columns

# Formula to find Missing Value
df.isnull().sum()

# Formula to find duplicate data
df.duplicated().sum()

# Data Cleaning

df

#The formula to remove Rp inside Price column
df['Price']= df['Price'].str.replace('Rp', '')
#The formula for removing the word Terjual from the Product sold column
df['Product sold']= df['Product sold'].str.replace('terjual', '')

#Formula to change the word rb to thousand(000) and Formula to remove the + sign in the Product sold column
df['Product sold']= df['Product sold'].str.replace('rb', '000')
df['Product sold']= df['Product sold'].str.replace('+', '')

#The formula to convert None to non values
df['Product sold'] = df['Product sold'].replace('None', pd.NA)

#The formula for removing spaces in the Product sold column
df['Product sold']= df['Product sold'].str.replace(' ', '')

#The formula to convert non values to 0 and change the data type from object to integer
df['Product sold'] = df['Product sold'].fillna(0).astype(int)

#The formula to convert non values to 0 and change the data type from object to float
df['Product rating']= df['Product rating'].fillna(0).astype(float)

# Formula to display Dataframe
df

#The formula to change the data type from object to float
df['Price']= df['Price'].astype(float)

# Formula to check the information within dataframe
df.info()

# Formula to drop Unnamed Column
df.drop(columns='Unnamed: 0', inplace=True)

# Formula to check the information within dataframe
df.info()

# Formula to export to csv file
df.to_csv('Seblak_tokped_clean.csv')
